+++ plugins/ubuntu2.sh
@@ -39,7 +39,12 @@ elif [ $1 = copy ];then
 		cp -R ubuntu2/casper multicd-working/boot/ubuntu2 #Live system
 		cp -R ubuntu2/preseed multicd-working/boot/ubuntu2
 		# Fix the isolinux.cfg
-		cp ubuntu2/isolinux/text.cfg multicd-working/boot/ubuntu2/ubuntu2.cfg
+		if [ -f ubuntu2/isolinux/text.cfg ];then
+			cp ubuntu2/isolinux/text.cfg multicd-working/boot/ubuntu2/ubuntu2.cfg
+		fi
+		if [ -f ubuntu2/isolinux/txt.cfg ];then
+			cp ubuntu2/isolinux/txt.cfg multicd-working/boot/ubuntu2/ubuntu2.cfg
+		fi
 		sed -i 's@default live@default menu.c32@g' multicd-working/boot/ubuntu2/ubuntu2.cfg
 		sed -i 's@file=/cdrom/preseed/@file=/cdrom/boot/ubuntu2/preseed/@g' multicd-working/boot/ubuntu2/ubuntu2.cfg
 		sed -i 's^initrd=/casper/^live-media-path=/boot/ubuntu2 ignore_uuid initrd=/boot/ubuntu2/^g' multicd-working/boot/ubuntu2/ubuntu2.cfg
